+++ StarWars/Player.cs
@@ -16,7 +16,7 @@ namespace StarWars
 
         public LaserHandler LaserHandler { get => laserHandler; }
 
-        public Player(Texture2D texture, int hitboxX, int hitboxY, int speed, int hitpoints, Texture2D laserTexture):base(texture, hitboxX, hitboxY, speed)
+        public Player(Texture2D texture, int hitboxX, int hitboxY, float speed, int hitpoints, Texture2D laserTexture):base(texture, hitboxX, hitboxY, speed)
         {
             //Set the start position
             Position = new Vector2((Game1.WindowWidth / 2) - (Hitbox.Width / 2), Game1.WindowHeight - Hitbox.Height - (Game1.WindowHeight * 0.05f));
